Rename internal to pkg
