+++ test/testutils/validators.go
@@ -32,7 +32,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	leaderworkerset "sigs.k8s.io/leader-worker-set/api/leaderworkerset/v1"
-	"sigs.k8s.io/leader-worker-set/internal/commonutils"
+	"sigs.k8s.io/leader-worker-set/pkg/commonutils"
 )
 
 const (
